Changes made for checkout

+++ src/main/java/com/onlinemart/model/ProductShoppingCart.java
@@ -23,16 +23,16 @@ public class ProductShoppingCart implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
-    private int quantity;
+    private Long quantity;
     
     @OneToOne
     private Product product;
     
-    public int getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
